basic materials

+++ src/Defs.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-constexpr const double kInfinity = std::numeric_limits<double>::max();
+constexpr const float kInfinity = std::numeric_limits<float>::max();
 using color = glm::u8vec3;
-using vec3 = glm::dvec3;
+using vec3 = glm::vec3;
