+++ src/cpu_raytrace/Interval.hpp
@@ -4,11 +4,11 @@
 namespace raytrace2::cpu {
 
 struct Interval {
-  double min{-kInfinity}, max{kInfinity};
-  [[nodiscard]] double Size() const { return max - min; }
+  float min{-kInfinity}, max{kInfinity};
+  [[nodiscard]] float Size() const { return max - min; }
 
-  [[nodiscard]] bool Contains(double x) const { return min <= x && x <= max; }
-  [[nodiscard]] bool Surrounds(double x) const { return min < x && x < max; }
+  [[nodiscard]] bool Contains(float x) const { return min <= x && x <= max; }
+  [[nodiscard]] bool Surrounds(float x) const { return min < x && x < max; }
 
   static const Interval kEmpty;
   static const Interval kUniverse;
