+++ src/cpu_raytrace/Ray.hpp
@@ -3,9 +3,9 @@
 namespace raytrace2::cpu {
 
 struct Ray {
-  [[nodiscard]] inline glm::dvec3 At(double t) const { return origin + direction * t; }
-  glm::dvec3 origin;
-  glm::dvec3 direction;
+  [[nodiscard]] inline glm::dvec3 At(float t) const { return origin + direction * t; }
+  glm::vec3 origin;
+  glm::vec3 direction;
 };
 
 }  // namespace raytrace2::cpu
